Issue #130, #131, #132: Created SingleNoteFragment that displays an individual note and its references, and refactored the layout file accordingly.
This is done in order to replace the dialog showing the single note. This needs to be done in order to open a new fragment from the single note, in order to edit or comment on the note. With a dialog, it would be impossible to navigate between all the pages. Additionally this allows us to make the single note page fullscreen.